backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
SQLAlchemy==2.0.23
Alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
boto3==1.34.18
openai==1.109.1
langchain==0.3.7
langchain-openai==0.2.8
langgraph==0.2.45
httpx==0.27.2
tiktoken==0.12.07
